New UI

+++ src/main/java/com/springapp/mvc/footballStats/controller/GameController.java
@@ -9,16 +9,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
+import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.view.RedirectView;
 
 import java.security.Principal;
 import java.util.List;
 
+@SessionAttributes({"players", "plays", "playresults"})
 @Controller
 public class GameController {
 
@@ -35,7 +35,10 @@ public class GameController {
 
     private Game game;
     @RequestMapping(value="/game/create", method=RequestMethod.GET)
-    public ModelAndView addGamePage(){
+    public ModelAndView addGamePage(Principal principal){
+        if(principal==null){
+            return new ModelAndView("login");
+        }
         ModelAndView modelAndView = new ModelAndView("createGame");
         modelAndView.addObject("game", new Game());
         List<Game> games = gameService.getGames();
@@ -44,8 +47,22 @@ public class GameController {
         return modelAndView;
     }
 
+    @RequestMapping(value="/game/delete/{id}", method=RequestMethod.GET)
+    public String deleteGame(@PathVariable Integer id, ModelMap model, Principal principal) {
+        if(principal==null){
+            return "login";
+        }
+        gameService.deleteGame(id);
+
+        return "redirect:../create";
+    }
+
     @RequestMapping(value="/game/edit/{id}", method=RequestMethod.GET)
-    public ModelAndView updateGame(@PathVariable Integer id, Principal principal){
+    public ModelAndView updateGame(@PathVariable Integer id,  Principal principal){
+        if(principal==null){
+            return new ModelAndView("login");
+        }
+
         final String currentUser = principal.getName();
         ModelAndView modelAndView = new ModelAndView("redirect:/game/playresult/add");
         game = gameService.getGame(id);
@@ -58,6 +75,9 @@ public class GameController {
 
     @RequestMapping(value="/game/create", method=RequestMethod.POST)
     public ModelAndView addGame(@ModelAttribute("Game")Game game, Principal principal){
+        if(principal==null){
+            return new ModelAndView("login");
+        }
         final String currentUser = principal.getName();
         game.setUser_Id(currentUser);
         ModelAndView modelAndView = new ModelAndView("redirect:/game/playresult/add");
@@ -72,34 +92,37 @@ public class GameController {
 
 
     @RequestMapping(value="/game/playresult/add", method=RequestMethod.GET)
-    public ModelAndView addPlayResultPage(){
-        if(game.getId()==null){
-            ModelAndView modelAndView = new ModelAndView("redirect:game/create");
-            return modelAndView;
+    public String addPlayResultPage(ModelMap model, Principal principal){
+        if(principal==null){
+            return "login";
         }
-        List<Player> players = playerService.getPlayers();
-        List<Play> plays = playbookService.getPlays();
-        ModelAndView modelAndView = new ModelAndView("recordPlay");
+        if(game.getId()==null){
+            return "createGame";
+        }
+        if(model.get("players")==null){
+            List<Player> players = playerService.getPlayers();
+            List<Play> plays = playbookService.getPlays();
+            model.addAttribute("plays", plays);
+            model.addAttribute("players", players);
+        }
         List<PlayResult> playResults = playResultService.getPlayResults(game.getId());
-        modelAndView.addObject("playresult", new PlayResult());
-        modelAndView.addObject("playresults", playResults);
-        modelAndView.addObject("plays", plays);
-        modelAndView.addObject("players", players);
+        model.addAttribute("playresult", new PlayResult());
+        model.addAttribute("playresults", playResults);
 
-        return modelAndView;
+        return "recordPlay";
     }
 
     @RequestMapping(value="/game/playresult/add", method=RequestMethod.POST)
     @Transactional
     public ModelAndView addPlayResult(@ModelAttribute("playresult") PlayResult playResult, BindingResult result, Principal principal){
+        if(principal==null){
+            return new ModelAndView("login");
+        }
         final String currentUser = principal.getName();
         playResult.setUser_Id(currentUser);
         playResult.setGame_Id(game.getId());
         playResult.setOpponent(game.getOpponent());
         playResult.setDate(game.getDate());
-
-        List<Player> players = playerService.getPlayers();
-        List<Play> plays = playbookService.getPlays();
         System.out.println(playResult.getPlay_Id());
 
         playResult.setCarrier(playerService.getPlayer(playResult.getCarrier_Id()).getName());
@@ -112,26 +135,28 @@ public class GameController {
         List<PlayResult> playResults = playResultService.getPlayResults(game.getId());
         modelAndView.addObject("playresult", new PlayResult());
         modelAndView.addObject("playresults", playResults);
-        modelAndView.addObject("plays", plays);
-        modelAndView.addObject("players", players);
 
         return modelAndView;
     }
 
     @RequestMapping(value="/game/playresult/list")
-    public ModelAndView listOfplays() {
+    public ModelAndView listOfplays(Principal principal) {
+        if(principal==null){
+            return new ModelAndView("login");
+        }
         ModelAndView modelAndView = new ModelAndView("list-of-playResults");
 
         List<PlayResult> playResults = playResultService.getPlayResults(game.getId());
-        List<Player> players = playerService.getPlayers();
-        modelAndView.addObject("players",players);
         modelAndView.addObject("playResults", playResults);
 
         return modelAndView;
     }
 
     @RequestMapping(value="/game/playresult/edit/{id}", method=RequestMethod.GET)
-    public ModelAndView editplayPage(@PathVariable Integer id) {
+    public ModelAndView editplayPage(@PathVariable Integer id, Principal principal) {
+        if(principal==null){
+            return new ModelAndView("login");
+        }
         ModelAndView modelAndView = new ModelAndView("edit-play-form");
         PlayResult playResult = playResultService.getPlayResult(id);
         modelAndView.addObject("playResult",playResult);
@@ -139,7 +164,10 @@ public class GameController {
     }
 
     @RequestMapping(value="/game/playresult/edit/{id}", method=RequestMethod.POST)
-    public ModelAndView editingplay(@ModelAttribute PlayResult playResult, @PathVariable Integer id) {
+    public ModelAndView editingplay(@ModelAttribute PlayResult playResult, @PathVariable Integer id, Principal principal) {
+        if(principal==null){
+            return new ModelAndView("login");
+        }
 
         ModelAndView modelAndView = new ModelAndView("login");
 
@@ -151,19 +179,22 @@ public class GameController {
         return modelAndView;
     }
 
-    @RequestMapping(value="/playresult/delete/{id}", method=RequestMethod.GET)
-    public ModelAndView deleteplay(@PathVariable Integer id) {
-        ModelAndView modelAndView = new ModelAndView("recordPlay");
+    @RequestMapping(value="/game/playresult/delete/{id}", method=RequestMethod.GET)
+    public String deleteplay(@PathVariable Integer id, ModelMap model, Principal principal) {
+        if(principal==null){
+            return "login";
+        }
         playResultService.deletePlayResult(id);
-        List<Player> players = playerService.getPlayers();
-        List<Play> plays = playbookService.getPlays();
-        List<PlayResult> playResults = playResultService.getPlayResults(game.getId());
-        modelAndView.addObject("playresult", new PlayResult());
-        modelAndView.addObject("playresults", playResults);
-        modelAndView.addObject("plays", plays);
-        modelAndView.addObject("players", players);
-        String message = "play was successfully deleted.";
-        modelAndView.addObject("message", message);
-        return modelAndView;
+
+        List<PlayResult> playResults = (List<PlayResult>) model.get("playresults");
+        for(int i = 0; i < playResults.size(); i++){
+            if(playResults.get(i).getId()==id){
+                playResults.remove(i);
+            }
+        }
+
+        model.addAttribute("playresult", new PlayResult());
+        model.addAttribute("playresults", playResults);
+        return "redirect:../add";
     }
 }
